OrePrefixes finished

+++ src/main/java/gregtech/api/material/type/DustMaterial.java
@@ -14,7 +14,7 @@ public class DustMaterial extends FluidMaterial {
 
     public static final class MatFlags {
 
-        public static final int GENERATE_ORE = createFlag(11);
+        public static final long GENERATE_ORE = createFlag(11);
 
         /**
          * Generate a plate for this material
@@ -22,31 +22,31 @@ public class DustMaterial extends FluidMaterial {
          * If it's metal material, bending machine recipes will be generated
          * If block is found, cutting machine recipe will be also generated
          */
-        public static final int GENERATE_PLATE = createFlag(12);
+        public static final long GENERATE_PLATE = createFlag(12);
 
         /**
          * Add to material if it cannot be worked by any other means, than smashing or smelting. This is used for coated Materials.
          */
-        public static final int NO_WORKING = createFlag(13);
+        public static final long NO_WORKING = createFlag(13);
         /**
          * Add to material if it cannot be used for regular Metal working techniques since it is not possible to bend it.
          */
-        public static final int NO_SMASHING = createFlag(14);
+        public static final long NO_SMASHING = createFlag(14);
 
         /**
          * Add to material if it's impossible to smelt it
          */
-        public static final int NO_SMELTING = createFlag(15);
+        public static final long NO_SMELTING = createFlag(15);
 
         /**
          * Add to material if it is outputting less in an Induction Smelter.
          */
-        public static final int INDUCTION_SMELTING_LOW_OUTPUT = createFlag(16);
+        public static final long INDUCTION_SMELTING_LOW_OUTPUT = createFlag(16);
 
         /**
          * Add to material if it melts into fluid (and it will also generate fluid for this material)
          */
-        public static final int SMELT_INTO_FLUID = createFlag(17);
+        public static final long SMELT_INTO_FLUID = createFlag(17);
 
     }
 
@@ -86,11 +86,11 @@ public class DustMaterial extends FluidMaterial {
      */
     public DustMaterial separatedOnto;
 
-    public DustMaterial(int metaItemSubId, String name, String defaultLocalName, int materialRGB, MaterialIconSet materialIconSet, ImmutableList<MaterialStack> materialComponents, int materialGenerationFlags, Element element) {
+    public DustMaterial(int metaItemSubId, String name, String defaultLocalName, int materialRGB, MaterialIconSet materialIconSet, ImmutableList<MaterialStack> materialComponents, long materialGenerationFlags, Element element) {
         super(metaItemSubId, name, defaultLocalName, materialRGB, materialIconSet, materialComponents, materialGenerationFlags, element);
     }
 
-    public DustMaterial(int metaItemSubId, String name, String defaultLocalName, int materialRGB, MaterialIconSet materialIconSet, ImmutableList<MaterialStack> materialComponents, int materialGenerationFlags) {
+    public DustMaterial(int metaItemSubId, String name, String defaultLocalName, int materialRGB, MaterialIconSet materialIconSet, ImmutableList<MaterialStack> materialComponents, long materialGenerationFlags) {
         super(metaItemSubId, name, defaultLocalName, materialRGB, materialIconSet, materialComponents, materialGenerationFlags, null);
     }
